bot/pyproject.toml
[project]
name = "auto-ads-service-bot"
version = "1.0.0"
description = "Bot searcher by LLM с Function Calling"
readme = "README.md"
requires-python = ">=3.13"
authors = [
    {name = "Anton Lukisha", email = "anton.lukisha.dev@gmail.com"},
]
license = {text = "MIT"}
dependencies = [
    "python-telegram-bot>=20.0",
    "python-dotenv>=1.0.0",
    "sqlalchemy>=2.0.0",
    "asyncpg>=0.29.0",
    "orjson>=3.11.7",
    "openai>=2.26.0"
]

[project.scripts]
bot = "src.bot:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = [
    "httpx>=0.28.1",
    "black>=26.1.0",
    "isort>=8.0.1",
    "flake8>=7.3.0",
    "mypy>=1.19.1",
    "ipython>=9.10.0",
    "types-requests>=2.32.4.20260107"
]

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.black]
line-length = 100
target-version = ["py313"]

[tool.isort]
profile = "black"
line_length = 100

[tool.mypy]
python_version = "3.13"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true